use rover/genesis root public key if not specfied
The onap parameter file can include a public key install in the
onap vm for access.  If the user does not provide a key then
the local account on the rover/genesis node where the install
is running will be used to make sure there is some access
available to the vm.

Signed-off-by: davidplunkett <dp7642@att.com>
Change-Id: I4b1a8070d9ef8da99c254d5ad8fa7a4aa6de2d71